requirements.txt
# Core ML Libraries
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
evaluate>=0.4.0
scikit-learn>=1.3.0

# Explainability
lime>=0.2.0.1
shap>=0.41.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Web Interface
streamlit>=1.25.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Configuration
PyYAML>=6.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Optional: GPU support
# torch-audio>=2.0.0
# torchvision>=0.15.0
